address feedback - move getMaxPushOutVolume to shared-data
