+++ app/src/organisms/ODD/QuickTransferFlow/utils/index.ts
@@ -8,4 +8,3 @@ export { getSelectedWellCount } from './getSelectedWellCount'
 export { getVolumeRange } from './getVolumeRange'
 export { getPipetteName } from './getPipetteName'
 export { getMaxUiFlowRate } from './getMaxUiFlowRate'
-export { getMaxPushOutVolume } from './getMaxPushOutVolume'
