+++ protocol-designer/src/utils/__tests__/utils.test.ts
@@ -5,7 +5,6 @@ import { fixture96Plate } from '@opentrons/shared-data'
 import {
   getAllLabwareIdsOfCertainURIOnStack,
   getMaxConditioningVolume,
-  getMaxPushOutVolume,
   removeOpentronsPhrases,
 } from '..'
 
@@ -62,66 +61,6 @@ describe('removeOpentronsPhrases', () => {
     const input = 'Eppendorf epT.I.P.S. Tip Rack is long'
     const expectedOutput = 'epT.I.P.S. Tip Rack is long'
     expect(removeOpentronsPhrases(input)).toBe(expectedOutput)
-  })
-})
-
-describe('getMaxPushOutVolume', () => {
-  const mockPipetteSpec: any = {
-    plungerPositionsConfigurations: {
-      default: {
-        bottom: 10,
-        blowout: 15,
-      },
-      lowVolumeDefault: {
-        bottom: 5,
-        blowout: 15,
-      },
-    },
-    shaftULperMM: 0.8,
-    liquids: {
-      default: {
-        maxVolume: 50,
-        minVolume: 5,
-      },
-      lowVolumeDefault: {
-        maxVolume: 30,
-        minVolume: 1,
-      },
-    },
-  }
-
-  it('should calculate correct push out volume for default volume configuration ', () => {
-    const result = getMaxPushOutVolume(100, mockPipetteSpec)
-
-    expect(result).toBe(4)
-  })
-
-  it('should calculate correct push out volume for low volume configuration ', () => {
-    const result = getMaxPushOutVolume(4, mockPipetteSpec)
-
-    expect(result).toBe(8)
-  })
-
-  it('should calculate pushout volume for low volume configuration with no low volume mode properties', () => {
-    const pipetteSpecNoLowVolume: any = {
-      plungerPositionsConfigurations: {
-        default: {
-          bottom: 10,
-          blowout: 15,
-        },
-      },
-      shaftULperMM: 0.8,
-      liquids: {
-        default: {
-          maxVolume: 50,
-          minVolume: 5,
-        },
-      },
-    }
-
-    const result = getMaxPushOutVolume(50, pipetteSpecNoLowVolume)
-
-    expect(result).toBe(4)
   })
 })
